+++ packages/expo-dev-launcher/ios/Tests/EXDevLauncherURLHelperTests.swift
@@ -5,9 +5,9 @@ import XCTest
 @testable import EXDevLauncher
 
 class EXDevLauncherURLHelperTests: XCTestCase {
-  
+
   let encodedUrlString = "http%3A%2F%2Flocalhost%3A8081"
-  
+
   func testIsDevLauncherURL() {
     let defaultUrl = "scheme://expo-development-client"
     XCTAssertTrue(EXDevLauncherURLHelper.isDevLauncherURL(URL(string: defaultUrl)))
@@ -22,28 +22,32 @@ class EXDevLauncherURLHelperTests: XCTestCase {
 
     let actual2 = EXDevLauncherURLHelper.replaceEXPScheme(URL(string: "http://expo-development-client/?url=http%3A%2F%2Flocalhost%3A8081")!, to: "scheme")
     XCTAssertEqual(URL(string: "http://expo-development-client/?url=http%3A%2F%2Flocalhost%3A8081"), actual2)
+
+    // should not crash if provided URL does not include scheme
+    let actual3 = EXDevLauncherURLHelper.replaceEXPScheme(URL(string: "192.168.0.12:19000")!, to: "scheme")
+    XCTAssertEqual(URL(string: "192.168.0.12:19000"), actual3)
   }
-  
+
   func testDevLauncherUrls() {
     // dev-client scheme with valid url param -> loadApp with specified url param
     expectDevLauncherUrlToEqual(input:"scheme://expo-development-client/?url=http%3A%2F%2Flocalhost%3A8081%2Findex.bundle%3Fplatform%3Dios%26dev%3Dtrue",
                          expected:"http://localhost:8081/index.bundle?platform=ios&dev=true")
-    
+
     // non-dev-client scheme with valid url param -> defer loading to loaded app
     expectDevLauncherUrlToEqual(input: "scheme://not-dev-client/?url=\(encodedUrlString)",
                                 expected: "scheme://not-dev-client/?url=\(encodedUrlString)")
 
   }
-  
+
   func testDevLauncherUrlQueryParams() {
     let url = "scheme://expo-development-client/?url=http%3A%2F%2Flocalhost%3A8081&updateMessage=123"
     let devLauncherUrl = EXDevLauncherUrl(URL(string:url)!)
     let queryParams = devLauncherUrl.queryParams
-    
+
     XCTAssertEqual(queryParams["updateMessage"], "123")
     XCTAssertEqual(queryParams["url"], "http://localhost:8081")
   }
-  
+
   //  HELPER
   func expectDevLauncherUrlToEqual(input: String, expected: String) {
     let devLauncherUrl = EXDevLauncherUrl(URL(string:input)!)